refactor(gateway): extract LazyTransportContext to gateway layer
LazyTransportContext + ReactToolContext lived in agent/tools/react.ts
because react was the first consumer that needed to call back into the
active transport. The mechanism is generic — any future transport-aware
tool (reply, send_typing, edit_message, ...) needs the same pattern,
and "ReactToolContext" was a misleading name for what is just
{ getTransport(): Transport | undefined }.

- New file src/gateway/transport-context.ts holds:
  - TransportContext interface (renamed from ReactToolContext)
  - LazyTransportContext class
- react.ts is now a pure consumer: imports TransportContext, defines
  no transport plumbing of its own
- Update 5 import sites + react.test.ts to point at gateway
